Fix typescript formating via gq
There is something in the base nvim config that sets a `formatexpr` for
typescript files that is preferred by nvim over what I set in the
`formatprg` option so it needs to be cleared.

However the way it was being being cleared seemed to be overrided by the
nvim base config. Scheduling the setting has fixed it for now.